Factor out route-util.ts.
